+++ tests/testthat/test-lockfile.R
@@ -69,7 +69,7 @@ test_that("we create lockfile from a manifest automatically when no lockfile fou
   expect_true(file.exists(file.path(project, "renv.lock")))
 
   unlink(project, recursive = TRUE)
-  
+
 })
 
 test_that("the Requirements field is read as character", {
@@ -121,5 +121,27 @@ test_that("lockfile APIs can be used", {
 
   # check that it's the same
   expect_equal(lockfile, lockfile_read())
+
+})
+
+test_that("lockfiles with UTF-8 contents can be written, read", {
+
+  author <- enc2utf8("cr\u00e8me br\u00fbl\u00e9e")
+  Encoding(author) <- "UTF-8"
+
+  lockfile <- list(
+    Packages = list(
+      example = list(
+        Package = "example",
+        Version = "1.0.0",
+        Author  = author
+      )
+    )
+  )
+
+  path <- renv_scope_tempfile("renv-lockfile-")
+  renv_lockfile_write(lockfile, file = path)
+  actual <- renv_lockfile_read(path)
+  expect_identical(unclass(actual), lockfile)
 
 })