+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# renv (development version)
+
 # renv 1.1.1
 
 * Fixed an issue where very long lockfiles could be truncated on write. (#2089)