+++ NEWS.md
@@ -1,12 +1,17 @@
 
 # renv (development version)
 
+* Fixed an issue where lockfiles containing unicode characters were not
+  correctly written or read with versions of R (< 4.2.0) on Windows. (#2091)
+
+
 # renv 1.1.1
 
 * Fixed an issue where very long lockfiles could be truncated on write. (#2089)
 
 * Fixed an issue where `renv` inadvertently made web requests in
   tests during `R CMD check`.
+
 
 # renv 1.1.0