+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# renv (development version)
+# renv 1.1.2
 
 * `renv` gains the `sysreqs()` function, which can be used to query the system
   packages required by a set of R packages. Functionality is currently available